put toggle theme and isBlackTheme logic in the Navbar
